+++ package.json
@@ -46,6 +46,7 @@
     "@changesets/cli": "2.17.0",
     "@commitlint/cli": "13.1.0",
     "@commitlint/config-conventional": "13.1.0",
+    "@google-cloud/bigquery": "5.9.0",
     "@types/jest": "27.0.1",
     "@types/pg": "8.6.1",
     "@jest/types": "27.1.1",
