+++ src/pg/index.ts
@@ -1,0 +1,3 @@
+export type { IPostgresInterval } from './interfaces';
+export { intervalStringValue, intervalToMilliseconds } from './interval';
+export { query } from './query';
