+++ src/pg/interval.ts
